Rewrite functional tests for new expected behavior - Expect lots of failures
